Tarea 2

+++ app/Http/Controllers/ProductController.php
@@ -2,6 +2,8 @@
 
 namespace App\Http\Controllers;
 
+use App\Http\Requests\StoreProductRequest;
+use App\Http\Requests\UpdateProductRequest;
 use App\Product;
 use Illuminate\Http\Request;
 use Illuminate\Http\Response;
@@ -27,7 +29,7 @@ class ProductController extends Controller
         'status' => Response::HTTP_OK]);
         }
         else {
-            return response()->json(NULL, 404);
+            return response()->json(NULL, 200);
         }
     }
 
@@ -44,11 +46,12 @@ class ProductController extends Controller
     /**
      * Store a newly created resource in storage.
      *
-     * @param  \Illuminate\Http\Request  $request
+     * @param  \App\Http\StoreProductRequest $request
      * @return \Illuminate\Http\Response
      */
-    public function store(Request $request)
+    public function store(StoreProductRequest $request)
     {
+        $validate = $request->validated();
          // Create a new product
         $product = Product::create($request->all());
 
@@ -71,7 +74,8 @@ class ProductController extends Controller
             return response()->json($product,200);
         }
         else {
-            return response()->json($product,404);
+            $error = ['errors' => ['code' => 'Error-2', 'title' => 'ID does not exist']];
+            return response()->json($error,404);
         }
         
     }
@@ -90,24 +94,23 @@ class ProductController extends Controller
     /**
      * Update the specified resource in storage.
      *
-     * @param  \Illuminate\Http\Request  $request
+     * @param  \App\Http\Request\UpdateProductRequest  $request
      * @param  \App\Product  $product
      * @return \Illuminate\Http\Response
      */
-    public function update($id)
+    public function update(UpdateProductRequest $request, $id)
     {
-        //
-        $data = $this->request->all();
+        $validate = $request->validated();
+
         $product = $this->product->find($id);
 
         if (!is_null($product)) {
-            $product->name = $data['name'];
-            $product->price = $data['price'];
-            $product->save();
+            $product->update($request->all());
             return response()->json($product, 200);
         }
         else{
-            return response()->json($product, 404);
+            $error = ['errors' => ['code' => 'Error-2', 'title' => 'ID does not exist']];
+            return response()->json($error, 404);
         }   
     }
 
@@ -122,8 +125,13 @@ class ProductController extends Controller
         //
         $product = $this->product->find($id);
 
-        $product->delete();
-
-        return response()->json(['status' => Response::HTTP_OK]);
+        if(!is_null($product)){
+            $product->delete();
+            return response()->json(NULL,204);
+        }
+        else {
+            $error = ['errors' => ['code' => 'Error-2', 'title' => 'ID does not exist']];
+            return response()->json($error,404);
+        }
     }
 }
